Compatibility: Avoid "E706: Variable type mismatch for: l:newExtension" in Vim 7.4.1099
Apparently, matchlist() covered up some implementation sloppiness by implicitly converting the List to a String, but Vim 7.4.1099 doesn't like the change in variable type from List to String (Vim 7.4.1689 is fine with it, though).
Explicitly use two variables for clarity: The initial is l:newExtensions of type List (though only one extension should be in there, as it doesn't make sense to pass multiple ones), and that gets translated through ingo#cmdargs#file#FilterFileOptionsAndCommandsToEscaped() if a:OptionParser is set (still to a List), and convert before passing to matchlist(), and assign the scalar result to l:newExtension.